Update to latest whl, move folder
