+++ previews/experimental-fragment/redirect/redirect_app.py
@@ -1,0 +1,5 @@
+import streamlit as st
+
+st.warning("This app has moved!")
+
+st.markdown("### [Open the app](https://st-experimental-fragment.streamlit.app/)")
