+++ previews/experimental-fragment/requirements.txt
@@ -1,0 +1,3 @@
+# This requirements.txt is just for Community Cloud
+# Download and install the whl file locally for your own testing!
+./previews/experimental-fragment/streamlit-1.32.2-py2.py3-none-any.whl
