--- previews/partial-reruns/requirements.txt
@@ -1,3 +1,0 @@
-# This requirements.txt is just for Community Cloud
-# Download and install the whl file locally for your own testing!
-./previews/partial-reruns/streamlit-1.32.0-py2.py3-none-any.whl
